Very basic organization
Set up some of the basic modules and classes needed for this project,
although most of them are minus some fairly critical bits to make them
actually be useful. Also added some additional detail to the readme to
explain some further ideas I've had about how to make this more than
just a room walk -- can support a variety of interesting other things
now as well, I think, in a mechanism that should be not very hard to
write and still be interesting to author.