Work on logic room flow, etc.
Started fleshing out a bit more of the way that we expect the server to
handle room transitions. Still kind of sketching this at the moment.
Also started filling in a little bit of idea on how to do the login flow
(again just as a sketch at the moment).