Revamp expected HTTP flow
Rearranged things so that we support some GET and some POST bits, and I
think we'll primarily use the latter for user actions during the game.
Still has some bits that aren't the greatest in that flow, but on the
other hand, this is a toy -- it will be just fine. Reloads are about the
only real weird thing there, and we could maybe redirect all POSTs into
GETs or something weird like that if we decide we really need something
along those lines.

Also dup'd the example into game so that we have a more realistic
starting point for really making something interesting there.