WIP login flow
Getting a bunch of pieces in place for the login and add user flows.
Kinda silly to do this first, I suppose, but I need to have these bits
in place so that I can actually run a user through for testing once I
start in on the room logic, so there you have it. Still a lot of
untested and unfinished code at this point, but want a safety checkin.