WIP login - done, maybe?
Added in the final (?) link up bits for how we handle a new session when
logging in or creating a new user. So theoretically we should now be
able to do something interesting when a user logs in. Maybe. Need to do
a bunch of testing next, although I plan to do a short detour to make a
data directory argument for running the server and such so that I can do
consistent/easy testing without needing a million commit bits sitting
around.