Login/Create flows working
Ran some tests and of course found a whole pile of things that were
utterly broken. Now seems to be working properly, so I can create a new
user and then log in as the user and go to the "room" placeholder that
we currently have. Checkpointing before I remove some debug stuff that
shouldn't be necessary any longer.