WIP room entry changes
Established a basic format for how to do session variable changes on
entry to a room. Updated the example with a few uses of this. I'm not
currently certain that I actually want the addrand operation; it might
be cleaner to instead have a mechanism to do add/set using variables
instead of constants, and thus use the setrand to make a var and then
add it to the contents, etc. But at any rate, this now allows for a
reasonable amount of dynamic content to happen.

The session currently doesn't save when changes are applied, which I
need to fix.