Room exit conditions
Now have a syntax and runtime for handling conditions associated with
exits. The "var" version of the comparison ops is actually pretty
decent, although I kinda think maybe I should consider making all ops be
numeric and then autodetect that the author is using vars instead of
values. But whatever, this at least works. I suspect I could use a
similar thing to make the changes system work with var/var setups
instead of var/value setups pretty easily, now that I've thought of how
it should flow in the condition case.

I think the next big bit to tackle is going to be the common
reset/settings bits, since right now once you finish the "game" in the
example, that account is basically stuck forever in that state.