Added daily plan, etc.
Fleshed out some additional bits on the project plan, and added a new
daily plan section with the plan for the following day. Most likely all
of those items will not be accomplished (it's a lot!) but it helps to
have a picture of the course that we're trying to cover.