Add logging and update daily plan
There's now some quasi-useful boilerplate messages that go to the output
log for the project, such that you can see some of what is going on when
the player takes actions. Expectation is that intern will want to expand
on this model a bit to get more insight during development.

Additionally, daily plan has been updated for next items on the project
plan. As always, expectation is that there is more to do than can
actually be done in a single day's session, but hopefully direction will
prove helpful in making progress.