Added hints for continued work
This hopefully will at least suggest where/how some things could be done
to make the grid room into a reality. Given the amount of work that's
there this may be a bit of a stretch, so we may want to reconsider what
should be done for the day's work.